Splitt filer for ventetid og venteperiode dataklasser
