+++ src/main/kotlin/no/nav/helse/flex/syketilfelle/ventetid/Venteperiode.kt
@@ -1,0 +1,19 @@
+package no.nav.helse.flex.syketilfelle.ventetid
+
+import no.nav.helse.flex.syketilfelle.sykmelding.domain.SykmeldingKafkaMessage
+import java.time.LocalDate
+
+data class VenteperiodeRequest(
+    val tilleggsopplysninger: Tilleggsopplysninger? = null,
+    val sykmeldingKafkaMessage: SykmeldingKafkaMessage? = null,
+    val harForsikring: Boolean = false,
+)
+
+data class VenteperiodeResponse(
+    val venteperiode: Venteperiode?,
+)
+
+data class Venteperiode(
+    val fom: LocalDate,
+    val tom: LocalDate,
+)
