--- src/main/kotlin/no/nav/helse/flex/syketilfelle/ventetid/VenteperiodeResponse.kt
@@ -1,12 +1,0 @@
-package no.nav.helse.flex.syketilfelle.ventetid
-
-import java.time.LocalDate
-
-data class VenteperiodeResponse(
-    val venteperiode: Venteperiode?,
-)
-
-data class Venteperiode(
-    val fom: LocalDate,
-    val tom: LocalDate,
-)
